pipeline working without move-pods
