+++ nifi-operator/kuttl-test.yaml
@@ -3,4 +3,4 @@ kind: TestSuite
 testDirs:
 - ./tests/kuttl
 startKIND: false
-timeout: 6000
+timeout: 600
