+++ nifi-operator/tests/kuttl/smoke/03-assert.yaml
@@ -1,0 +1,9 @@
+#---
+#apiVersion: apps/v1
+#kind: StatefulSet
+#metadata:
+#  name: test-nifi-node-default
+#status:
+#  readyReplicas: 3
+#  replicas: 3
+
