+++ nifi-operator/tests/kuttl/smoke/03-move-pod.sh
