+++ nifi-operator/tests/kuttl/smoke/03-move-pod.yaml
@@ -1,0 +1,5 @@
+#---
+#apiVersion: kuttl.dev/v1beta1
+#kind: TestStep
+#commands:
+#  - command: ./03-move-pod.sh
