+++ nifi-operator/tests/kuttl/smoke/04-assert.yaml
@@ -2,13 +2,14 @@
 apiVersion: kuttl.dev/v1beta1
 kind: TestAssert
 metadata:
-  name: install-nifi
-timeout: 1200
+  name: install-test-nifi
+timeout: 300
 ---
 apiVersion: apps/v1
 kind: StatefulSet
 metadata:
-  name: test-nifi-node-default
+  name: test-nifi
 status:
-  readyReplicas: 2
-  replicas: 2
+  readyReplicas: 1
+  replicas: 1
+
