--- nifi-operator/tests/kuttl/smoke/04-checknodes-scaleup.yaml
@@ -1,3 +1,0 @@
----
-apiVersion: kuttl.dev/v1beta1
-kind: TestStep
