+++ nifi-operator/tests/kuttl/smoke/05-assert.yaml
@@ -1,9 +1,7 @@
 ---
-apiVersion: apps/v1
-kind: StatefulSet
+apiVersion: kuttl.dev/v1beta1
+kind: TestAssert
 metadata:
-  name: test-nifi-node-default
-status:
-  readyReplicas: 3
-  replicas: 3
-
+  name: test-nifi
+commands:
+  - script: kubectl exec -n $NAMESPACE test-nifi-0 -- python /tmp/test_nifi.py -u admin -p supersecretpassword -n $NAMESPACE
