--- nifi-operator/tests/kuttl/smoke/05-move-pod.yaml
@@ -1,5 +1,0 @@
----
-apiVersion: kuttl.dev/v1beta1
-kind: TestStep
-commands:
-  - command: ./05-move-pod.sh
