+++ nifi-operator/tests/kuttl/smoke/05-prepare-test-nifi.yaml
@@ -1,0 +1,7 @@
+apiVersion: kuttl.dev/v1beta1
+kind: TestStep
+commands:
+  - script: kubectl cp -n $NAMESPACE ./test_nifi.py  test-nifi-0:/tmp
+  - script: kubectl cp -n $NAMESPACE ./cacert.pem  test-nifi-0:/tmp
+  - script: kubectl cp -n $NAMESPACE ./requirements.txt test-nifi-0:/tmp
+  - script: kubectl exec -n $NAMESPACE test-nifi-0 -- pip install --user -r /tmp/requirements.txt
