--- nifi-operator/tests/kuttl/smoke/07-assert.yaml
@@ -1,39 +1,0 @@
----
-apiVersion: v1
-kind: Event
-reason: Started
-source:
-  component: kubelet
-involvedObject:
-  apiVersion: v1
-  kind: Pod
-  name: test-nifi-node-default-0
----
-apiVersion: v1
-kind: Event
-reason: Started
-source:
-  component: kubelet
-involvedObject:
-  apiVersion: v1
-  kind: Pod
-  name: test-nifi-node-default-1
----
-apiVersion: v1
-kind: Event
-reason: Started
-source:
-  component: kubelet
-involvedObject:
-  apiVersion: v1
-  kind: Pod
-  name: test-nifi-node-default-2
----
-apiVersion: apps/v1
-kind: StatefulSet
-metadata:
-  name: test-nifi-node-default
-status:
-  readyReplicas: 3
-  replicas: 3
-
