+++ nifi-operator/tests/kuttl/smoke/requirements.txt
@@ -1,0 +1,1 @@
+requests==2.27.1
