--- nifi-operator/tests/script/script/01-assert.yaml
@@ -1,8 +1,0 @@
----
-apiVersion: apps/v1
-kind: StatefulSet
-metadata:
-  name: test-nifi-node-default
-status:
-  readyReplicas: 3
-  replicas: 3
