--- nifi-operator/tests/script/script/01-move-pod.yaml
@@ -1,5 +1,0 @@
----
-apiVersion: kuttl.dev/v1beta1
-kind: TestStep
-commands:
-  - command: ./01-move-pod.sh
